Refactor class inheritance to use vtds-base provided base classes
