+++ vtds_application_mock/private/config.py
@@ -28,7 +28,7 @@ from . import CONFIG_DIR
 
 
 # pylint: disable=too-few-public-methods
-class PrivateBaseConfig(BaseConfiguration):
+class BaseConfig(BaseConfiguration):
     """BaseConfig class presents operations on the base configuration
     of the layer to callers.
 
@@ -37,4 +37,4 @@ class PrivateBaseConfig(BaseConfiguration):
         """Constructor
 
         """
-        super().__init__("mock application", CONFIG_DIR)
+        BaseConfiguration.__init__(self, "mock application", CONFIG_DIR)
